+++ Classes/Utility/ExtensionUtility.php
@@ -13,7 +13,7 @@ class ExtensionUtility {
    * Gets an identifier from an extension key.
    *
    * @param string $extensionKey The extension key
-   * @return string The extension identifier
+   * @return string The identifier
    */
   public static function identifier(string $extensionKey): string {
     $identifier = GeneralUtility::underscoredToUpperCamelCase($extensionKey);
@@ -23,62 +23,18 @@ class ExtensionUtility {
   }
 
   /**
-   * Alias for the `identifier` function.
-   *
-   * @param string $extensionKey The extension key
-   * @return string The extension identifier
-   */
-  public static function extensionIdentifier(string $extensionKey): string {
-    return self::identifier($extensionKey);
-  }
-
-  /**
-   * Alias for the `extensionIdentifier` function.
-   *
-   * @param string $extensionKey The extension key
-   * @return string The extension identifier
-   */
-  public static function getExtensionIdentifier(string $extensionKey): string {
-    return self::extensionIdentifier($extensionKey);
-  }
-
-  /**
    * Gets a signature from a namespace and extension key.
    *
    * @param string $namespace The namespace
    * @param string $extensionKey The extension key
    * @param string $separator The optional separator, defaults to `.`
-   * @return string The extension signature
+   * @return string The signature
    */
   public static function signature(string $namespace, string $extensionKey, string $separator = '.'): string {
     $namespace    = GeneralUtility::underscoredToUpperCamelCase($namespace);
     $extensionKey = GeneralUtility::underscoredToUpperCamelCase($extensionKey);
 
     return "${namespace}${separator}${extensionKey}";
-  }
-
-  /**
-   * Alias for the `signature` function.
-   *
-   * @param string $namespace The namespace
-   * @param string $extensionKey The extension key
-   * @param string $separator The optional separator, defaults to `.`
-   * @return string The extension signature
-   */
-  public static function extensionSignature(string $namespace, string $extensionKey, string $separator = '.'): string {
-    return self::signature($namespace, $extensionKey, $separator);
-  }
-
-  /**
-   * Alias for the `extensionSignature` function.
-   *
-   * @param string $namespace The namespace
-   * @param string $extensionKey The extension key
-   * @param string $separator The optional separator, defaults to `.`
-   * @return string The extension signature
-   */
-  public static function getExtensionSignature(string $namespace, string $extensionKey, string $separator = '.'): string {
-    return self::extensionSignature($namespace, $extensionKey, $separator);
   }
 
   /**
@@ -94,19 +50,6 @@ class ExtensionUtility {
     $languageFolder = self::languageFolder($extensionKey, $prefix);
 
     return "${languageFolder}/${fileName}${separator}";
-  }
-
-  /**
-   * Alias for the `locallang` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional file name of the locallang file, defaults to `locallang.xlf`
-   * @param string $prefix The optional prefix, defaults to `LLL:EXT:`
-   * @param string $separator The optional separator, defaults to `:`
-   * @return string The locallang file
-   */
-  public static function getLocallang(string $extensionKey, string $fileName = 'locallang.xlf', string $prefix = 'LLL:EXT:', string $separator = ':'): string {
-    return self::locallang($extensionKey, $fileName, $prefix, $separator);
   }
 
   /**
@@ -134,17 +77,6 @@ class ExtensionUtility {
   }
 
   /**
-   * Alias for the `configurationFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `FILE:EXT:`
-   * @return string The configuration folder
-   */
-  public static function getConfigurationFolder(string $extensionKey, string $prefix = 'FILE:EXT:'): string {
-    return self::configurationFolder($extensionKey, $prefix);
-  }
-
-  /**
    * Gets the FlexForms folder from an extension key.
    *
    * @param string $extensionKey The extension key
@@ -158,17 +90,6 @@ class ExtensionUtility {
   }
 
   /**
-   * Alias for the `flexFormsFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `FILE:EXT:`
-   * @return string The FlexForms folder
-   */
-  public static function getFlexFormsFolder(string $extensionKey, string $prefix = 'FILE:EXT:'): string {
-    return self::flexFormsFolder($extensionKey, $prefix);
-  }
-
-  /**
    * Gets the resources folder from an extension key.
    *
    * @param string $extensionKey The extension key
@@ -177,17 +98,6 @@ class ExtensionUtility {
    */
   public static function resourcesFolder(string $extensionKey, string $prefix = 'EXT:'): string {
     return "${prefix}${extensionKey}/Resources";
-  }
-
-  /**
-   * Alias for the `resourcesFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `EXT:`
-   * @return string The resources folder
-   */
-  public static function getResourcesFolder(string $extensionKey, string $prefix = 'EXT:'): string {
-    return self::resourcesFolder($extensionKey, $prefix);
   }
 
   /**
@@ -204,17 +114,6 @@ class ExtensionUtility {
   }
 
   /**
-   * Alias for the `privateFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `EXT:`
-   * @return string The private folder
-   */
-  public static function getPrivateFolder(string $extensionKey, string $prefix = 'EXT:'): string {
-    return self::privateFolder($extensionKey, $prefix);
-  }
-
-  /**
    * Gets the language folder from an extension key.
    *
    * @param string $extensionKey The extension key
@@ -228,17 +127,6 @@ class ExtensionUtility {
   }
 
   /**
-   * Alias for the `languageFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `EXT:`
-   * @return string The language folder
-   */
-  public static function getLanguageFolder(string $extensionKey, string $prefix = 'EXT:'): string {
-    return self::languageFolder($extensionKey, $prefix);
-  }
-
-  /**
    * Gets the locallang folder from an extension key.
    *
    * Alias for the `languageFolder` function with `LLL:EXT:` as prefix.
@@ -249,17 +137,6 @@ class ExtensionUtility {
    */
   public static function locallangFolder(string $extensionKey, string $prefix = 'LLL:EXT:'): string {
     return self::languageFolder($extensionKey, $prefix);
-  }
-
-  /**
-   * Alias for the `locallangFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `LLL:EXT:`
-   * @return string The locallang folder
-   */
-  public static function getLocallangFolder(string $extensionKey, string $prefix = 'LLL:EXT:'): string {
-    return self::locallangFolder($extensionKey, $prefix);
   }
 
   /**
@@ -287,17 +164,6 @@ class ExtensionUtility {
   }
 
   /**
-   * Alias for the `publicFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `EXT:`
-   * @return string The public folder
-   */
-  public static function getPublicFolder(string $extensionKey, string $prefix = 'EXT:'): string {
-    return self::publicFolder($extensionKey, $prefix);
-  }
-
-  /**
    * Gets the assets folder from an extension key.
    *
    * @param string $extensionKey The extension key
@@ -308,17 +174,6 @@ class ExtensionUtility {
     $publicFolder = self::publicFolder($extensionKey, $prefix);
 
     return "${publicFolder}/Assets";
-  }
-
-  /**
-   * Alias for the `assetsFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `EXT:`
-   * @return string The assets folder
-   */
-  public static function getAssetsFolder(string $extensionKey, string $prefix = 'EXT:'): string {
-    return self::assetsFolder($extensionKey, $prefix);
   }
 
   /**
@@ -335,17 +190,6 @@ class ExtensionUtility {
   }
 
   /**
-   * Alias for the `iconsFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `EXT:`
-   * @return string The icons folder
-   */
-  public static function getIconsFolder(string $extensionKey, string $prefix = 'EXT:'): string {
-    return self::iconsFolder($extensionKey, $prefix);
-  }
-
-  /**
    * Gets the placeholders folder from an extension key.
    *
    * @param string $extensionKey The extension key
@@ -359,17 +203,6 @@ class ExtensionUtility {
   }
 
   /**
-   * Alias for the `placeholdersFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `EXT:`
-   * @return string The placeholders folder
-   */
-  public static function getPlaceholdersFolder(string $extensionKey, string $prefix = 'EXT:'): string {
-    return self::placeholdersFolder($extensionKey, $prefix);
-  }
-
-  /**
    * Gets the samples folder from an extension key.
    *
    * @param string $extensionKey The extension key
@@ -380,16 +213,5 @@ class ExtensionUtility {
     $publicFolder = self::publicFolder($extensionKey, $prefix);
 
     return "${publicFolder}/Samples";
-  }
-
-  /**
-   * Alias for the `samplesFolder` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $prefix The optional prefix, defaults to `EXT:`
-   * @return string The samples folder
-   */
-  public static function getSamplesFolder(string $extensionKey, string $prefix = 'EXT:'): string {
-    return self::samplesFolder($extensionKey, $prefix);
   }
 }