+++ Classes/Utility/IconUtility.php
@@ -24,17 +24,6 @@ class IconUtility {
   }
 
   /**
-   * Alias for the `iconIdentifier` function.
-   *
-   * @param string $iconKey The icon key
-   * @param string $separator The optional separator, defaults to `_`
-   * @return string The icon identifier
-   */
-  public static function getIconIdentifier(string $iconKey, string $separator = '_'): string {
-    return self::iconIdentifier($iconKey, $separator);
-  }
-
-  /**
    * Gets an icon signature from an extension key and icon identifier.
    *
    * @param string $extensionKey The extension key
@@ -46,17 +35,5 @@ class IconUtility {
     $iconSignature = mb_strtolower("${extensionKey}${separator}${iconIdentifier}");
 
     return $iconSignature;
-  }
-
-  /**
-   * Alias for the `iconSignature` function.
-   *
-   * @param string $extensionKey The extension key
-   * @param string $iconIdentifier The icon identifier
-   * @param string $separator The optional separator, defaults to `-`
-   * @return string The icon signature
-   */
-  public static function getIconSignature(string $extensionKey, string $iconIdentifier, string $separator = '-'): string {
-    return self::iconSignature($extensionKey, $iconIdentifier, $separator);
   }
 }